push

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: naessgui <naessgui@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/23 10:24:36 by naessgui          #+#    #+#             */
-/*   Updated: 2024/11/07 21:35:47 by naessgui         ###   ########.fr       */
+/*   Updated: 2024/11/15 18:57:02 by naessgui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
@@ -24,8 +24,9 @@ size_t	ft_strlen(const char *s)
 	return (i);
 }
 // #include <stdio.h>
-// int main(){
-// 	char s[] = "nadaessghiar";
-// 	printf ("%zu", ft_strlen(s));
+// int	main(void)
+// {
+// 	char s[] = "nada essghiar";
+// 	printf("%zu", ft_strlen(s));
 // 	return (0);
 // }
